Продолжаю работу над реализацией основных механик

+++ Assets/__Scripts/Damage.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 
 public struct Damage 
 {
-    public int attackDmg;
+    public float attackDmg;
 }
